Added confirmed field to database and back-end model, added constants file for better structure
