+++ backend/routes/admins.js
@@ -7,17 +7,18 @@ const authUser = require('../middleware/authUser');
 const adminController = require('../controllers/adminController');
 const authController = require('../controllers/authController');
 const authJwt = require('../middleware/authJwt');
+const constants = require("../utils/constants");
 
 router.use(express.json());
 router.use(authUser.setUser);
 
 //Admin test page
-router.get('/', authUser.verifyUser, authUser.verifyRole(authUser.ROLE.ADMIN), (req, res) => {
+router.get('/', authUser.verifyUser, authUser.verifyRole(constants.ROLE.ADMIN), (req, res) => {
     res.send("If you can access this then you are an admin !");
 });
 
-// Update user roles as admin
-router.put('/updateRole', authUser.verifyUser, authUser.verifyRole(authUser.ROLE.ADMIN), (req, res) => {
+//Update user roles as admin
+router.put('/updateRole', authUser.verifyUser, authUser.verifyRole(constants.ROLE.ADMIN), (req, res) => {
     adminController.updateRole(req,res);
 })
