+++ backend/utils/constants.js
@@ -1,0 +1,11 @@
+const ROLE = {
+    ADMIN: 'Admin',
+    DOCTOR: 'Doctor',
+    HEALTH_OFFICIAL: 'HealthOfficial',
+    IMMIGRATION_OFFICER: 'ImmigrationOfficer',
+    PATIENT: 'Patient',
+}
+
+module.exports = {
+    ROLE
+}
